requirements.txt
# FinSIght Platform Requirements

# Core dependencies
agno>=2.1.8
fastapi>=0.119.0
uvicorn>=0.38.0
pydantic>=2.12.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0

# Market data
yfinance>=0.2.40
requests>=2.31.0

# Database
alembic>=1.13.0

# Optional: Production deployment
gunicorn>=21.2.0
redis>=5.0.0

# Optional: Advanced features
apscheduler>=3.10.4  # Background task scheduler (production-grade)
pytz>=2024.1  # Timezone support for scheduler
celery>=5.3.0  # Alternative: For background tasks
twilio>=8.0.0  # For SMS notifications
sendgrid>=6.11.0  # For email
pandas>=2.2.0  # For data analysis
numpy>=1.26.0  # For calculations

# Crypto wallet integration
web3>=7.0.0  # Web3 Python library for EVM interactions
eth-account>=0.13.0  # Ethereum account management and SIWE
PyJWT>=2.8.0  # JWT token generation and verification

# Development
pytest>=8.0.0
black>=24.0.0
flake8>=7.0.0
